MAINT: rename vis*.py files to qt_*.py
