+++ hnn/qt_dipole.py
