+++ hnn/qt_spec.py
