+++ hnn/qt_thread.py
